move geoinfo to separate app
