+++ claim/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
-from claim.models import Claim, Organization, InCharge,\
-    OrganizationType, Layer, Polygon
+from claim.models import Claim, InCharge,\
+    Organization, OrganizationType
 
 
 class ClaimAdmin(admin.ModelAdmin):
@@ -23,15 +23,7 @@ class InChargeAdmin(admin.ModelAdmin):
     list_filter = ('name', 'email')
 
 
-class PolygonAdmin(admin.ModelAdmin):
-    list_display = ('polygon_id', 'layer')
-    search_fields = ('olygon_id',)
-    list_filter = ('layer',)
-
-
 admin.site.register(Claim, ClaimAdmin)
 admin.site.register(Organization, OrganizationAdmin)
 admin.site.register(InCharge, InChargeAdmin)
 admin.site.register(OrganizationType)
-admin.site.register(Layer)
-admin.site.register(Polygon, PolygonAdmin)
