+++ claim/management/commands/initiate_db.py
@@ -3,7 +3,8 @@ import json
 from django.core.management.base import BaseCommand
 from django.utils.translation import ugettext as _
 
-from claim.models import OrganizationType, Organization, Layer, Polygon
+from claim.models import OrganizationType, Organization
+from geoinfo.models import Layer, Polygon
 from utils.common import read_map
