+++ corruption_tracker/settings.py
@@ -45,6 +45,7 @@ INSTALLED_APPS = (
 
     'leaflet',
     'claim',
+    'geoinfo',
 )
 
 MIDDLEWARE_CLASSES = (
